Fix case sensitivity
Windows is not case sensitive, so before the capital b in
headerBarStyles didn't matter. Since Docker runs using wsl it becomes
linux, which is case sensitive, so running as a docker container then
becomes an issue. Need to be careful of this. Also remove volumes from
docker-compose.dev.yml because we only need one container